[Core] use uint32 for tickTreeRoot

+++ src/core/contracts/base/TickStructure.sol
@@ -12,7 +12,7 @@ abstract contract TickStructure is AlgebraPoolBase {
   using TickManagement for mapping(int24 => TickManagement.Tick);
   using TickTree for mapping(int16 => uint256);
 
-  uint64 internal tickTreeRoot; // The root of bitmap search tree
+  uint32 internal tickTreeRoot; // The root of bitmap search tree
   mapping(int16 => uint256) internal tickSecondLayer; // The second layer bitmap search tree
 
   // the leaves of the tree are stored in `tickTable`
@@ -33,7 +33,7 @@ abstract contract TickStructure is AlgebraPoolBase {
     int24 prevInitializedTick,
     bool remove
   ) internal override returns (int24 newPrevInitializedTick) {
-    uint64 oldTickTreeRoot = tickTreeRoot;
+    uint32 oldTickTreeRoot = tickTreeRoot;
 
     int24 prevTick;
     if (remove) {
@@ -52,7 +52,7 @@ abstract contract TickStructure is AlgebraPoolBase {
       ticks.insertTick(tick, prevTick, nextTick);
     }
 
-    uint64 newTickTreeRoot = tickTable.toggleTick(tickSecondLayer, tick, oldTickTreeRoot);
+    uint32 newTickTreeRoot = tickTable.toggleTick(tickSecondLayer, tick, oldTickTreeRoot);
     if (newTickTreeRoot != oldTickTreeRoot) tickTreeRoot = newTickTreeRoot;
     return prevInitializedTick;
   }
